+++ commerce/delivery/src/main/java/ru/yandex/practicum/service/DeliveryService.java
@@ -3,12 +3,13 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.dto.order.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
     DeliveryDto planDelivery(DeliveryDto deliveryDto);
 
-    Double deliveryCost(OrderDto orderDto);
+    BigDecimal deliveryCost(OrderDto orderDto);
 
     void deliveryPicked(UUID orderId);
