+++ commerce/order/src/main/java/ru/yandex/practicum/model/Order.java
@@ -17,6 +17,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import ru.yandex.practicum.dto.order.OrderState;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -52,9 +53,9 @@ public class Order {
     private Double deliveryVolume;
     private Boolean fragile;
     @Column(name = "total_price")
-    private Double totalPrice;
+    private BigDecimal totalPrice;
     @Column(name = "delivery_price")
-    private Double deliveryPrice;
+    private BigDecimal deliveryPrice;
     @Column(name = "product_price")
-    private Double productPrice;
+    private BigDecimal productPrice;
 }
